+++ events/aggregator/vote.go
@@ -1,8 +1,6 @@
 package aggregator
 
 import (
-	"encoding/json"
-
 	"github.com/goverland-labs/platform-events/events"
 )
 
@@ -21,15 +19,4 @@ type VotePayload struct {
 
 type VotesPayload []VotePayload
 
-type VotesHandler func(VotesPayload) error
-
-func (h VotesHandler) RawHandler() events.RawMessageHandler {
-	return func(raw []byte) error {
-		var d VotesPayload
-		if err := json.Unmarshal(raw, &d); err != nil {
-			return err
-		}
-
-		return h(d)
-	}
-}
+type VotesHandler = events.Handler[VotesPayload]