+++ events/core/dao.go
@@ -1,8 +1,6 @@
 package core
 
 import (
-	"encoding/json"
-
 	"github.com/google/uuid"
 
 	"github.com/goverland-labs/platform-events/events"
@@ -59,15 +57,4 @@ type DaoPayload struct {
 	ActiveSince    *int              `json:"active_since,omitempty"`
 }
 
-type DaoHandler func(payload DaoPayload) error
-
-func (h DaoHandler) RawHandler() events.RawMessageHandler {
-	return func(raw []byte) error {
-		var d DaoPayload
-		if err := json.Unmarshal(raw, &d); err != nil {
-			return err
-		}
-
-		return h(d)
-	}
-}
+type DaoHandler = events.Handler[DaoPayload]